+++ test/reduce.jl
@@ -2,7 +2,7 @@
 
 # fold(l|r) & mapfold(l|r)
 @test foldl(+, Int64[]) === Int64(0) # In reference to issues #7465/#20144 (PR #20160)
-@test foldl(+, Int16[]) === Int32(0)
+@test foldl(+, Int16[]) === Int16(0) # In reference to issues #21536
 @test foldl(-, 1:5) == -13
 @test foldl(-, 10, 1:5) == -5
 
@@ -19,7 +19,7 @@
 @test Base.mapfoldl((x)-> x ⊻ true, |, false, [true false true false false]) == true
 
 @test foldr(+, Int64[]) === Int64(0) # In reference to issue #20144 (PR #20160)
-@test foldr(+, Int16[]) === Int32(0)
+@test foldr(+, Int16[]) === Int16(0) # In reference to issues #21536
 @test foldr(-, 1:5) == 3
 @test foldr(-, 10, 1:5) == -7
 @test foldr(+, [1]) == 1 # Issue #21493
@@ -29,7 +29,7 @@
 
 # reduce
 @test reduce(+, Int64[]) === Int64(0) # In reference to issue #20144 (PR #20160)
-@test reduce(+, Int16[]) === Int32(0)
+@test reduce(+, Int16[]) === Int16(0) # In reference to issues #21536
 @test reduce((x,y)->"($x+$y)", 9:11) == "((9+10)+11)"
 @test reduce(max, [8 6 7 5 3 0 9]) == 9
 @test reduce(+, 1000, 1:5) == (1000 + 1 + 2 + 3 + 4 + 5)
@@ -48,8 +48,8 @@
 @test mapreduce(abs2, +, Float64[]) === 0.0
 @test mapreduce(abs2, Base.scalarmax, Float64[]) === 0.0
 @test mapreduce(abs, max, Float64[]) === 0.0
-@test mapreduce(abs2, &, Float64[]) === true
-@test mapreduce(abs2, |, Float64[]) === false
+@test_throws ArgumentError mapreduce(abs2, &, Float64[])
+@test_throws ArgumentError mapreduce(abs2, |, Float64[])
 
 # mapreduce() type stability
 @test typeof(mapreduce(*, +, Int8[10])) ===
@@ -69,16 +69,28 @@
       typeof(mapreduce(abs, +, Float32[10, 11, 12, 13]))
 
 # sum
+@testset "sums promote to at least machine size" begin
+    @testset for T in [Int8, Int16, Int32]
+        @test sum(T[]) === Int(0)
+    end
+    @testset for T in [UInt8, UInt16, UInt32]
+        @test sum(T[]) === UInt(0)
+    end
+    @testset for T in [Int, Int64, Int128, UInt, UInt64, UInt128,
+                       Float16, Float32, Float64]
+        @test sum(T[]) === T(0)
+    end
+    @test sum(BigInt[]) == big(0) && sum(BigInt[]) isa BigInt
+end
 
-@test sum(Int8[]) === Int32(0)
-@test sum(Int[]) === Int(0)
-@test sum(Float64[]) === 0.0
+@test sum(Bool[]) === sum(Bool[false]) === sum(Bool[false, false]) === 0
+@test sum(Bool[true, false, true]) === 2
 
-@test sum(Int8(3)) === Int8(3)
+@test sum(Int8(3)) === Int(3)
 @test sum(3) === 3
 @test sum(3.0) === 3.0
 
-@test sum([Int8(3)]) === Int32(3)
+@test sum([Int8(3)]) === Int(3)
 @test sum([3]) === 3
 @test sum([3.0]) === 3.0
 
@@ -135,14 +147,22 @@ end
 @test sum_kbn([-0.0]) === -0.0
 @test sum_kbn([-0.0,-0.0]) === -0.0
 
+# check sum(abs, ...) for support of empty collections
+@testset "sum(abs, [])" begin
+    @test @inferred(sum(abs, Float64[])) === 0.0
+    @test @inferred(sum(abs, Int[])) === 0
+    @test @inferred(sum(abs, Set{Int}())) === 0
+    @test_throws MethodError sum(abs, Any[])
+end
+
 # prod
 
 @test prod(Int[]) === 1
-@test prod(Int8[]) === Int32(1)
+@test prod(Int8[]) === Int(1)
 @test prod(Float64[]) === 1.0
 
 @test prod([3]) === 3
-@test prod([Int8(3)]) === Int32(3)
+@test prod([Int8(3)]) === Int(3)
 @test prod([3.0]) === 3.0
 
 @test prod(z) === 120
@@ -157,8 +177,8 @@ end
 prod2(itr) = invoke(prod, Tuple{Any}, itr)
 @test prod(Int[]) === prod2(Int[]) === 1
 @test prod(Int[7]) === prod2(Int[7]) === 7
-@test typeof(prod(Int8[])) == typeof(prod(Int8[1])) == typeof(prod(Int8[1, 7])) == Int32
-@test typeof(prod2(Int8[])) == typeof(prod2(Int8[1])) == typeof(prod2(Int8[1 7])) == Int32
+@test typeof(prod(Int8[])) == typeof(prod(Int8[1])) == typeof(prod(Int8[1, 7])) == Int
+@test typeof(prod2(Int8[])) == typeof(prod2(Int8[1])) == typeof(prod2(Int8[1 7])) == Int
 
 # maximum & minimum & extrema
 
@@ -380,3 +400,7 @@ test18695(r) = sum( t^2 for t in r )
 
 # issue #21107
 @test foldr(-,2:2) == 2
+
+# test neutral element not picked incorrectly for &, |
+@test @inferred(foldl(&, Int[1])) === 1
+@test_throws ArgumentError foldl(&, Int[])